justdata/apps/lenderprofile/services/ffiec_client.py
#!/usr/bin/env python3
"""
FFIEC CRA Performance Evaluations Client
Scrapes CRA ratings and performance evaluations (no formal API).
"""

import requests
import logging
from typing import Optional, Dict, Any, List
from bs4 import BeautifulSoup

logger = logging.getLogger(__name__)


class FFIECClient:
    """
    Client for FFIEC CRA Performance Evaluations.
    
    Base URL: https://www.ffiec.gov/craadweb/
    Note: No formal API, requires web scraping
    """
    
    def __init__(self):
        """Initialize FFIEC client."""
        self.base_url = 'https://www.ffiec.gov'
        self.cra_url = f'{self.base_url}/craadweb/'
        self.timeout = 30
        self.user_agent = 'NCRC Lender Intelligence Platform (contact@ncrc.org)'
    
    def _get_headers(self) -> Dict[str, str]:
        """Get request headers."""
        return {'User-Agent': self.user_agent}
    
    def search_cra_evaluations(self, cert: str) -> List[Dict[str, Any]]:
        """
        Search for CRA performance evaluations by FDIC certificate number.
        
        Args:
            cert: FDIC certificate number
            
        Returns:
            List of CRA evaluations
        """
        try:
            # FFIEC CRA database search
            # This is a placeholder - actual implementation would scrape the site
            url = self.cra_url
            params = {'cert': cert}
            
            response = requests.get(url, params=params, headers=self._get_headers(), timeout=self.timeout)
            response.raise_for_status()
            
            # Parse HTML
            soup = BeautifulSoup(response.text, 'html.parser')
            evaluations = []
            
            # TODO: Implement actual scraping logic based on FFIEC site structure
            # This would extract:
            # - Exam date
            # - CRA rating (Outstanding, Satisfactory, Needs to Improve, Substantial Noncompliance)
            # - Test-level ratings (lending, investment, service)
            # - Performance evaluation PDF link
            # - Examiner findings
            
            return evaluations
            
        except requests.exceptions.RequestException as e:
            logger.error(f"FFIEC error searching for cert {cert}: {e}")
            return []

